Basic Routes Setup
Routes were setup for the root "/" (which gets redirected to "/webquiz", and a route to a login page "login".
This layout will change. The ideas was to get the framework set up to use the route, view, and controller files.
(This layout is following the locallibrary tutorial by medium.com )