HTTPS : proxy middleware added for WebSocket connections
WebSocket handshakes use the Upgrade header and bypasses Express's middleware. This hooks into the HTPS server and manually forwards the upgrade requests.
This in conjunction with the changes to the OHIF Viewer build changes for HTTPS gets rid of the CORS issues and there are no more WebSocket handshake errors.